Move NumbersFormat to the print module
