+++ crates/sncast/src/lib.rs
@@ -50,30 +50,6 @@ struct Account {
     class_hash: Option<String>,
 }
 
-#[derive(Debug, PartialEq, Eq, Copy, Clone)]
-pub enum NumbersFormat {
-    Default,
-    Decimal,
-    Hex,
-}
-
-impl NumbersFormat {
-    #[must_use]
-    pub fn from_flags(hex_format: bool, dec_format: bool) -> Self {
-        assert!(
-            !(hex_format && dec_format),
-            "Exclusivity should be validated by clap"
-        );
-        if hex_format {
-            NumbersFormat::Hex
-        } else if dec_format {
-            NumbersFormat::Decimal
-        } else {
-            NumbersFormat::Default
-        }
-    }
-}
-
 pub struct WaitForTx {
     pub wait: bool,
     pub timeout: u16,
